更新了一下Models里边的文件

+++ PixePixie/Models/PPPlayer.h
@@ -4,7 +4,7 @@
 {
     NSString * playerName;
     long playerStone;
-    int playerLevel;
+    long playerLevel;
     
     NSArray * playerFriends;
     NSArray * playerPixies;
